refactor(db) 8/13: extract db/world.rs (closes #254)
Move 7 world/spatial populators (appearance/fx/scripts/nodes/planet/spawn/
conversation), fxspec_fqn_from_path, find_planet_code_after_cc (re-exported
pub(crate) -- shared with conquest), the ConversationRefCounts struct, their
CREATE TABLEs, and 3 tests into db/world.rs. 240 tests + clippy -D warnings + fmt green.
